scripts/requirements.txt
Adafruit-Blinka==8.54.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-requests==4.1.9
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
board==1.0
loguru==0.7.3
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
sysv_ipc==1.1.0
typing_extensions==4.12.2
